fix setPower (and inline some more stuff)
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@657 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ light.cpp
@@ -650,9 +650,9 @@ void Light::PrepareMoversCustom()
 
    Material mtrl;
 
-   const float r = (m_d.m_color & 255) * (float)(1.0/255.0);
-   const float g = (m_d.m_color & 65280) * (float)(1.0/65280.0);
-   const float b = (m_d.m_color & 16711680) * (float)(1.0/16711680.0);
+   const float r = (float)(m_d.m_color & 255) * (float)(1.0/255.0);
+   const float g = (float)(m_d.m_color & 65280) * (float)(1.0/65280.0);
+   const float b = (float)(m_d.m_color & 16711680) * (float)(1.0/16711680.0);
 
    Pin3D * const ppin3d = &g_pplayer->m_pin3d;
    for(int i=0; i<2; i++)
@@ -857,9 +857,9 @@ void Light::RenderCustomMovers(const RenderDevice* _pd3dDevice)
 
    Material mtrl;
    
-   const float r = (m_d.m_color & 255) * (float)(1.0/255.0);
-   const float g = (m_d.m_color & 65280) * (float)(1.0/65280.0);
-   const float b = (m_d.m_color & 16711680) * (float)(1.0/16711680.0);
+   const float r = (float)(m_d.m_color & 255) * (float)(1.0/255.0);
+   const float g = (float)(m_d.m_color & 65280) * (float)(1.0/65280.0);
+   const float b = (float)(m_d.m_color & 16711680) * (float)(1.0/16711680.0);
 
    for (int i=0; i<2; i++)
    {
@@ -1009,9 +1009,9 @@ void Light::RenderMovers(const RenderDevice* _pd3dDevice)
       NumVideoBytes += 32*sizeof(Vertex3D);     
    }
 
-   const float r = (m_d.m_color & 255) * (float)(1.0/255.0);
-   const float g = (m_d.m_color & 65280) * (float)(1.0/65280.0);
-   const float b = (m_d.m_color & 16711680) * (float)(1.0/16711680.0);
+   const float r = (float)(m_d.m_color & 255) * (float)(1.0/255.0);
+   const float g = (float)(m_d.m_color & 65280) * (float)(1.0/65280.0);
+   const float b = (float)(m_d.m_color & 16711680) * (float)(1.0/16711680.0);
 
    Material mtrl;
    for (int i=0; i<2; i++)
